tests/runtime/backup/direct_source_coverage.py
#!/usr/bin/env python3
"""
Direct source code execution with coverage for rest_adapter.py.
This is a specialized approach that runs coverage directly on the source code.
"""
import os
import sys
import asyncio
import coverage
from unittest.mock import MagicMock

# Create the coverage object targeting our specific file
cov = coverage.Coverage(source=['runtime.rest_adapter'])

# Start coverage collection
cov.start()

# Execute the rest_adapter.py module directly
with open('/workspaces/copilot/runtime/rest_adapter.py', 'r') as f:
    code = compile(f.read(), '/workspaces/copilot/runtime/rest_adapter.py', 'exec')
    
    # Set up the mock context
    context = {
        '__name__': '__main__',  # Ensure the main block runs
        '__file__': '/workspaces/copilot/runtime/rest_adapter.py'
    }
    
    # Create mocks for all dependencies
    fastapi_mock = MagicMock()
    app_mock = MagicMock()
    fastapi_mock.FastAPI.return_value = app_mock
    
    # Create HTTPException class
    class MockHTTPException(Exception):
        def __init__(self, status_code=None, detail=None):
            self.status_code = status_code
            self.detail = detail
            super().__init__(f"HTTP {status_code}: {detail}")
    
    fastapi_mock.HTTPException = MockHTTPException
    
    # Mock all required modules
    sys.modules['fastapi'] = fastapi_mock
    sys.modules['fastapi.responses'] = MagicMock()
    sys.modules['runtime.dispatcher'] = MagicMock()
    sys.modules['uvicorn'] = MagicMock()
    
    # Execute the code
    exec(code, context)

# Stop coverage and report
cov.stop()
cov.save()

print("\n📊 Coverage Report for rest_adapter.py:")
cov.report(include=["runtime/rest_adapter.py"], show_missing=True)
cov.html_report(directory='htmlcov')
