pyproject.toml
[tool.poetry]
name = "azure-components-foundry"
version = "0.1.0"
description = "A foundry for creating reusable components for Azure cloud services"
authors = ["Your Organization <info@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
azure-identity = "^1.15.0"
azure-keyvault-secrets = "^4.7.0"
azure-storage-blob = "^12.19.0"
pydantic = "^2.5.0"
rpds-py = "^0.25.1"
azure-ai-ml = "^1.27.1"
# Runtime dependencies for unified container
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
azure-functions = "^1.18.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^4.1.0"
black = "^23.11.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
pytest-asyncio = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88